Giant Split of BedRequestDataService because it was too long
