+++ BedBrigade.Common/Constants/Defaults.cs
@@ -23,6 +23,7 @@
         public const string FailedDeliveryText = "Failed delivery";
         public const string SameDayScheduleText = "Same day schedule";
         public const int DefaultDistance = 999;
+        public const double AverageDaysInAMonth = 30.44; //This includes leap year
         
         public const string DefaultBedType = "Single Bed";
         //The reason why I have chosen 12 is because each row is four bootstrap columns
