+++ BedBrigade.Data/Services/IBedRequestDataService.cs
@@ -7,27 +7,11 @@ namespace BedBrigade.Data.Services
     {
         Task<ServiceResponse<List<BedRequest>>> GetAllForLocationAsync(int locationId);
         Task<ServiceResponse<List<BedRequest>>> GetBedRequestsForUser();
-        Task<ServiceResponse<List<string>>> GetDistinctEmail();
-        Task<ServiceResponse<List<string>>> GetDistinctEmailByLocation(int locationId);
-        Task<ServiceResponse<List<string>>> EmailsForNotReceivedABed(int locationId);
-        Task<ServiceResponse<List<string>>> EmailsForReceivedABed(int locationId);
-        Task<ServiceResponse<List<string>>> EmailsForSchedule(int locationId);
         Task<ServiceResponse<List<BedRequest>>> GetAllForLocationList(List<int> locationIds);
         Task<ServiceResponse<List<BedRequest>>> GetScheduledBedRequestsForLocation(int locationId);
-        Task<ServiceResponse<int>> SumBedsForNotReceived(int locationId);
-        Task<ServiceResponse<BedRequest>> GetByPhone(string phone);
-        Task<ServiceResponse<List<string>>> GetDistinctPhone();
-        Task<ServiceResponse<List<string>>> GetDistinctPhoneByLocation(int locationId);
-        Task<ServiceResponse<List<string>>> PhonesForNotReceivedABed(int locationId);
-        Task<ServiceResponse<List<string>>> PhonesForReceivedABed(int locationId);
-        Task<ServiceResponse<List<string>>> PhonesForSchedule(int locationId);
-        List<BedRequest> SortBedRequestClosestToAddress(List<BedRequest> bedRequests, int bedRequestId);
-        Task<ServiceResponse<BedRequest>> GetWaitingByEmail(string email);
-        Task<ServiceResponse<BedRequest>> GetWaitingByPhone(string phone);
         Task<int> MarkInvalidEmailForWaitingForBedRequest(List<string> emailList);
         Task<ServiceResponse<DateTime?>> NextDateEligibleForBedRequest(NewBedRequest bedRequest);
         Task<ServiceResponse<List<BedRequest>>> GetBedRequestsByUserAndStatus(List<BedRequestStatus> statuses);
         Task<ServiceResponse<List<BedRequest>>> GetAllForScheduleId(int scheduleId);
-        Task<ServiceResponse<List<BedRequest>>> GetReplacementBedRequests(BedRequest bedRequest);
     }
 }
