+++ BedBrigade.Data/Services/IBedRequestEstimatedWaitDataService.cs
@@ -1,0 +1,10 @@
+﻿using BedBrigade.Common.Enums;
+using BedBrigade.Common.Models;
+
+namespace BedBrigade.Data.Services
+{
+    public interface IBedRequestEstimatedWaitDataService : IRepository<BedRequest>
+    {
+        Task FillEstimatedWait(BedRequest bedRequest);
+    }
+}
